refactor(headless-future): remove endpoint client config indirection
Extract internal configuration reader helpers, reuse them in endpoint facades/loaders, and delete redundant endpoint-client-configuration wrapper and tests.
