+++ packages/headless-future/src/core/interface/api/conversation-endpoint/conversation-endpoint-facade.ts
@@ -5,9 +5,9 @@ import {
 } from '@/src/api/index.js';
 import {EndpointFacade} from '@/src/core/internal/api/base-facade/endpoint-facade.js';
 import {buildRequest} from '@/src/core/internal/api/base-facade/endpoint-facade-request-builder.js';
-import {buildEndpointClientConfiguration} from '@/src/core/internal/api/base-facade/endpoint-client-configuration.js';
 import {getEndpointContributorRegistry} from '@/src/core/internal/api/base-facade/endpoint-contributor-registry.js';
 import {conversationEndpointKey} from '@/src/core/internal/api/base-facade/endpoint-keys.js';
+import {readEndpointClientConfiguration} from '@/src/core/internal/configuration/configuration-reader.js';
 import {FullEngine} from '@/src/core/interface/engine/engine.js';
 import * as conversationEndpointMutators from './conversation-endpoint-mutators.js';
 import type {ConversationEndpointCallResult} from './conversation-endpoint-types.js';
@@ -56,7 +56,7 @@ export class ConversationEndpointFacade extends EndpointFacade<CoveoConversation
       ...this.getOrderedRequestContributors(),
     ]);
 
-    const clientConfiguration = buildEndpointClientConfiguration(engine);
+    const clientConfiguration = readEndpointClientConfiguration(engine);
 
     const result = await this.#client.call(finalRequest, clientConfiguration);
